chore: try to fix compilation error at CI

+++ src/runtime/object.cpp
@@ -74,7 +74,7 @@ extern "C" void lean_inc_ref_cold(lean_object * o) {
 }
 
 extern "C" void lean_inc_ref_n_cold(lean_object * o, unsigned n) {
-    std::atomic_fetch_sub_explicit(lean_get_rc_mt_addr(o), n, std::memory_order_relaxed);
+    std::atomic_fetch_sub_explicit(lean_get_rc_mt_addr(o), (int)n, std::memory_order_relaxed);
 }
 
 extern "C" size_t lean_object_byte_size(lean_object * o) {
